organize legacy test files
