+++ packages/contracts/test/randomizer/legacy/polyptychRandomizerV0.test.ts
@@ -19,13 +19,17 @@ import {
   deployCoreWithMinterFilter,
   mintProjectUntilRemaining,
   advanceEVMByTime,
-} from "../util/common";
-import { FOUR_WEEKS } from "../util/constants";
+} from "../../util/common";
+import { FOUR_WEEKS } from "../../util/constants";
+
+import { Logger } from "@ethersproject/logger";
+// hide nuisance logs about event overloading
+Logger.setLogLevel(Logger.levels.ERROR);
 
 /**
- * General Integration tests for Basic Randomizer V2 with V3 core.
+ * General Integration tests for Polyptych Randomizer V0 with V3 core.
  */
-describe("BasicRandomizerV2 w/V3 core", async function () {
+describe("PolyptychRandomizerV0 w/V3 core", async function () {
   async function _beforeEach() {
     let config: T_Config = {
       accounts: await getAccounts(),
@@ -40,13 +44,42 @@ describe("BasicRandomizerV2 w/V3 core", async function () {
       adminACL: config.adminACL,
     } = await deployCoreWithMinterFilter(
       config,
-      "GenArt721CoreV3",
-      "MinterFilterV1"
+      "GenArt721CoreV3_Engine",
+      "MinterFilterV1",
+      false,
+      undefined,
+      "BasicPolyptychRandomizerV0"
     ));
+
+    config.delegationRegistry = await deployAndGet(
+      config,
+      "DelegationRegistry",
+      []
+    );
 
     config.minter = await deployAndGet(config, "MinterSetPriceV2", [
       config.genArt721Core.address,
       config.minterFilter.address,
+    ]);
+
+    ({
+      genArt721Core: config.genArt721CorePolyptych,
+      minterFilter: config.minterFilterPolyptych,
+      randomizer: config.randomizerPolyptych,
+      adminACL: config.adminACLPolyptych,
+    } = await deployCoreWithMinterFilter(
+      config,
+      "GenArt721CoreV3_Engine",
+      "MinterFilterV1",
+      false,
+      undefined,
+      "BasicPolyptychRandomizerV0"
+    ));
+
+    config.minterPolyptych = await deployAndGet(config, "MinterPolyptychV0", [
+      config.genArt721CorePolyptych.address,
+      config.minterFilterPolyptych.address,
+      config.delegationRegistry.address,
     ]);
 
     // add project
@@ -57,6 +90,15 @@ describe("BasicRandomizerV2 w/V3 core", async function () {
       .connect(config.accounts.deployer)
       .toggleProjectIsActive(config.projectZero);
     await config.genArt721Core
+      .connect(config.accounts.artist)
+      .updateProjectMaxInvocations(config.projectZero, config.maxInvocations);
+    await config.genArt721CorePolyptych
+      .connect(config.accounts.deployer)
+      .addProject("name", config.accounts.artist.address);
+    await config.genArt721CorePolyptych
+      .connect(config.accounts.deployer)
+      .toggleProjectIsActive(config.projectZero);
+    await config.genArt721CorePolyptych
       .connect(config.accounts.artist)
       .updateProjectMaxInvocations(config.projectZero, config.maxInvocations);
 
@@ -70,14 +112,57 @@ describe("BasicRandomizerV2 w/V3 core", async function () {
     await config.minter
       .connect(config.accounts.artist)
       .updatePricePerTokenInWei(config.projectZero, 0);
+    await config.minterFilterPolyptych
+      .connect(config.accounts.deployer)
+      .addApprovedMinter(config.minterPolyptych.address);
+    await config.minterFilterPolyptych
+      .connect(config.accounts.deployer)
+      .setMinterForProject(config.projectZero, config.minterPolyptych.address);
+    await config.minterPolyptych
+      .connect(config.accounts.artist)
+      .updatePricePerTokenInWei(config.projectZero, 0);
     return config;
   }
+
+  describe("Polyptych project hash seed settings", function () {
+    it("only the artist can set a project as polyptych", async function () {
+      const config = await loadFixture(_beforeEach);
+      await config.randomizerPolyptych
+        .connect(config.accounts.artist)
+        .toggleProjectIsPolyptych(0);
+      await expectRevert(
+        config.randomizerPolyptych
+          .connect(config.accounts.deployer)
+          .toggleProjectIsPolyptych(0),
+        "Only Artist"
+      );
+    });
+    it("requires the hash seed setter contract to be configured", async function () {
+      const config = await loadFixture(_beforeEach);
+      await expectRevert(
+        config.randomizerPolyptych
+          .connect(config.accounts.deployer)
+          .setPolyptychHashSeed(0, "0x000000000000000000000000"),
+        "Only hashSeedSetterContract"
+      );
+      await config.randomizerPolyptych.setHashSeedSetterContract(
+        config.accounts.deployer.address
+      );
+      await config.randomizerPolyptych
+        .connect(config.accounts.deployer)
+        .setPolyptychHashSeed(0, "0x000000000000000000000000");
+    });
+  });
 
   describe("assignCoreAndRenounce", function () {
     it("does not allow non-owner to call", async function () {
       const config = await loadFixture(_beforeEach);
       // deploy new randomizer
-      const randomizer = await deployAndGet(config, "BasicRandomizerV2", []);
+      const randomizer = await deployAndGet(
+        config,
+        "BasicPolyptychRandomizerV0",
+        []
+      );
       // expect failure when non-owner calls renounce and assign core
       await expectRevert(
         randomizer
@@ -90,7 +175,11 @@ describe("BasicRandomizerV2 w/V3 core", async function () {
     it("allows owner to call", async function () {
       const config = await loadFixture(_beforeEach);
       // deploy new randomizer
-      const randomizer = await deployAndGet(config, "BasicRandomizerV2", []);
+      const randomizer = await deployAndGet(
+        config,
+        "BasicPolyptychRandomizerV0",
+        []
+      );
       // expect success when owner calls renounce and assign core
       await randomizer
         .connect(config.accounts.deployer)
@@ -100,7 +189,11 @@ describe("BasicRandomizerV2 w/V3 core", async function () {
     it("does not allow owner to call twice", async function () {
       const config = await loadFixture(_beforeEach);
       // deploy new randomizer
-      const randomizer = await deployAndGet(config, "BasicRandomizerV2", []);
+      const randomizer = await deployAndGet(
+        config,
+        "BasicPolyptychRandomizerV0",
+        []
+      );
       // expect failure when owner calls renounce and assign core twice
       await randomizer
         .connect(config.accounts.deployer)
@@ -166,6 +259,54 @@ describe("BasicRandomizerV2 w/V3 core", async function () {
       console.info(`projectZeroTokenZeroHash: ${projectZeroTokenZeroHash}`);
       console.info(`projectZeroTokenOneHash: ${projectZeroTokenOneHash}`);
     });
+
+    it("copies the token hash from a previous token as expected", async function () {
+      const config = await loadFixture(_beforeEach);
+      await config.minterPolyptych
+        .connect(config.accounts.deployer)
+        .registerNFTAddress(config.genArt721Core.address);
+      await config.minterPolyptych
+        .connect(config.accounts.artist)
+        .allowHoldersOfProjects(
+          config.projectZero,
+          [config.genArt721Core.address],
+          [config.projectZero]
+        );
+      await config.randomizerPolyptych
+        .connect(config.accounts.deployer)
+        .setHashSeedSetterContract(config.minterPolyptych.address);
+      await config.randomizerPolyptych
+        .connect(config.accounts.artist)
+        .toggleProjectIsPolyptych(0);
+
+      await expectRevert(
+        config.minterPolyptych
+          .connect(config.accounts.artist)
+          ["purchase(uint256,address,uint256)"](
+            config.projectZero,
+            config.genArt721Core.address,
+            0,
+            {
+              value: 0,
+            }
+          ),
+        "Cannot have an empty hash seed to copy."
+      );
+
+      await config.minter.connect(config.accounts.artist).purchase(0);
+
+      // mint a token
+      await config.minterPolyptych
+        .connect(config.accounts.artist)
+        ["purchase(uint256,address,uint256)"](
+          config.projectZero,
+          config.genArt721Core.address,
+          0,
+          {
+            value: 0,
+          }
+        );
+    });
   });
 
   describe("owner", function () {
@@ -179,7 +320,11 @@ describe("BasicRandomizerV2 w/V3 core", async function () {
     it("returns deployer prior to being configured and renounced", async function () {
       const config = await loadFixture(_beforeEach);
       // deploy new randomizer
-      const randomizer = await deployAndGet(config, "BasicRandomizerV2", []);
+      const randomizer = await deployAndGet(
+        config,
+        "BasicPolyptychRandomizerV0",
+        []
+      );
       // expect owner to be deployer prior to being configured and renounced
       expect(await randomizer.owner()).to.be.equal(
         config.accounts.deployer.address
